# swagger.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "FinVeda API",
                 "version":  "v1"
             },
    "paths":  {
...
                  "/api/v1/Installments/loan/{loanId}":  {
                                                             "get":  {
                                                                         "tags":  [
                                                                                      "Installments"
                                                                                  ],
                                                                         "parameters":  [
                                                                                            {
                                                                                                "name":  "loanId",
                                                                                                "in":  "path",
                                                                                                "required":  true,
                                                                                                "schema":  "@{type=string; format=uuid}"
                                                                                            }
                                                                                        ],
                                                                         "responses":  {
                                                                                           "200":  {
                                                                                                       "description":  "OK",
                                                                                                       "content":  "@{text/plain=; application/json=; text/json=}"
                                                                                                   }
                                                                                       }
                                                                     }
                                                         },
...
                  "/api/v1/LoanCases":  {
                                            "post":  {
                                                         "tags":  [
                                                                      "LoanCases"
                                                                  ],
                                                         "requestBody":  {
                                                                             "content":  {
                                                                                             "application/json":  "@{schema=}",
                                                                                             "text/json":  "@{schema=}",
                                                                                             "application/*+json":  "@{schema=}"
                                                                                         }
                                                                         },
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "OK"
                                                                                   }
                                                                       }
                                                     },
                                            "get":  {
                                                        "tags":  [
                                                                     "LoanCases"
                                                                 ],
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "OK",
                                                                                      "content":  "@{text/plain=; application/json=; text/json=}"
                                                                                  }
                                                                      }
                                                    }
                                        },
                  "/api/v1/LoanCases/{id}":  {
                                                 "get":  {
                                                             "tags":  [
                                                                          "LoanCases"
                                                                      ],
                                                             "parameters":  [
                                                                                {
                                                                                    "name":  "id",
                                                                                    "in":  "path",
                                                                                    "required":  true,
                                                                                    "schema":  "@{type=string; format=uuid}"
                                                                                }
                                                                            ],
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK",
                                                                                           "content":  "@{text/plain=; application/json=; text/json=}"
                                                                                       }
                                                                           }
                                                         }
                                             },
                  "/api/v1/LoanCases/{id}/approve":  {
                                                         "post":  {
                                                                      "tags":  [
                                                                                   "LoanCases"
                                                                               ],
                                                                      "parameters":  [
                                                                                         {
                                                                                             "name":  "id",
                                                                                             "in":  "path",
                                                                                             "required":  true,
                                                                                             "schema":  "@{type=string; format=uuid}"
                                                                                         }
                                                                                     ],
                                                                      "responses":  {
                                                                                        "200":  {
                                                                                                    "description":  "OK"
                                                                                                }
                                                                                    }
                                                                  }
                                                     },
                  "/api/v1/LoanCases/{id}/disburse":  {
                                                          "post":  {
                                                                       "tags":  [
                                                                                    "LoanCases"
                                                                                ],
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "id",
                                                                                              "in":  "path",
                                                                                              "required":  true,
                                                                                              "schema":  "@{type=string; format=uuid}"
                                                                                          }
                                                                                      ],
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "OK"
                                                                                                 }
                                                                                     }
                                                                   }
                                                      },
...
              },
    "components":  {
...
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "http",
                                                              "description":  "Enter \u0027Bearer\u0027 [space] and then your valid token.",
                                                              "scheme":  "Bearer",
                                                              "bearerFormat":  "JWT"
                                                          }
                                           }
                   },
    "security":  [
                     {
                         "Bearer":  [

                                    ]
                     }
                 ]
}